refactor(#38): 유저를 생성하는 로직 위치를 UserQueryService에서 UserService로 이동
